k8s: Split SharedResources into binary specific cells
Split SharedResources cell into multiple cells, one for each binary
relying on k8s resources (agent, operator and clustermesh-apiserver).
Each binary depends only on the specific resources that it actually
uses, and it customizes each resource events stream based on its own
specific needs.
