--- pkg/k8s/shared_resources.go
@@ -1,206 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-package k8s
-
-import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/fields"
-
-	"github.com/cilium/cilium/pkg/hive"
-	"github.com/cilium/cilium/pkg/hive/cell"
-	cilium_api_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
-	cilium_api_v2alpha1 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2alpha1"
-	"github.com/cilium/cilium/pkg/k8s/client"
-	"github.com/cilium/cilium/pkg/k8s/resource"
-	slim_corev1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
-	"github.com/cilium/cilium/pkg/k8s/utils"
-	nodeTypes "github.com/cilium/cilium/pkg/node/types"
-	"github.com/cilium/cilium/pkg/option"
-)
-
-var (
-	// SharedResourceCell provides a set of shared handles to Kubernetes resources used throughout the
-	// Cilium agent. Each of the resources share a client-go informer and backing store so we only
-	// have one watch API call for each resource kind and that we maintain only one copy of each object.
-	//
-	// See pkg/k8s/resource/resource.go for documentation on the Resource[T] type.
-	SharedResourcesCell = cell.Module(
-		"k8s-shared-resources",
-		"Shared Kubernetes resources",
-
-		cell.Provide(
-			serviceResource,
-			func(lc hive.Lifecycle, cs client.Clientset) (LocalNodeResource, error) {
-				return nodeResource(
-					lc, cs,
-					func(opts *metav1.ListOptions) {
-						opts.FieldSelector = fields.ParseSelectorOrDie("metadata.name=" + nodeTypes.GetName()).String()
-					},
-				)
-			},
-			func(lc hive.Lifecycle, cs client.Clientset) (LocalCiliumNodeResource, error) {
-				return ciliumNodeResource(
-					lc, cs,
-					func(opts *metav1.ListOptions) {
-						opts.FieldSelector = fields.ParseSelectorOrDie("metadata.name=" + nodeTypes.GetName()).String()
-					},
-				)
-			},
-			func(lc hive.Lifecycle, cs client.Clientset) (LocalPodResource, error) {
-				return podResource(
-					lc, cs,
-					func(opts *metav1.ListOptions) {
-						opts.FieldSelector = fields.ParseSelectorOrDie("spec.nodeName=" + nodeTypes.GetName()).String()
-					},
-				)
-			},
-			namespaceResource,
-			lbIPPoolsResource,
-			ciliumIdentityResource,
-			ciliumNetworkPolicyResource,
-			ciliumClusterwideNetworkPolicyResource,
-			ciliumCIDRGroupResource,
-		),
-	)
-)
-
-type SharedResources struct {
-	cell.In
-	LocalNode                        LocalNodeResource
-	LocalCiliumNode                  LocalCiliumNodeResource
-	LocalPods                        LocalPodResource
-	Services                         resource.Resource[*slim_corev1.Service]
-	Namespaces                       resource.Resource[*slim_corev1.Namespace]
-	LBIPPools                        resource.Resource[*cilium_api_v2alpha1.CiliumLoadBalancerIPPool]
-	Identities                       resource.Resource[*cilium_api_v2.CiliumIdentity]
-	CiliumNetworkPolicies            resource.Resource[*cilium_api_v2.CiliumNetworkPolicy]
-	CiliumClusterwideNetworkPolicies resource.Resource[*cilium_api_v2.CiliumClusterwideNetworkPolicy]
-	CIDRGroups                       resource.Resource[*cilium_api_v2alpha1.CiliumCIDRGroup]
-}
-
-// LocalPodResource is a resource.Resource[*slim_corev1.Pod] but one which will only stream updates for pod
-// objects scheduled on the node we are currently running on.
-type LocalPodResource resource.Resource[*slim_corev1.Pod]
-
-func podResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*slim_corev1.Pod], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*slim_corev1.PodList](cs.Slim().CoreV1().Pods("")),
-		opts...,
-	)
-	return resource.New[*slim_corev1.Pod](lc, lw), nil
-}
-
-func serviceResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*slim_corev1.Service], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	optsModifier, err := utils.GetServiceListOptionsModifier(option.Config)
-	if err != nil {
-		return nil, err
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*slim_corev1.ServiceList](cs.Slim().CoreV1().Services("")),
-		append(opts, optsModifier)...,
-	)
-	return resource.New[*slim_corev1.Service](lc, lw), nil
-}
-
-// LocalNodeResource is a resource.Resource[*corev1.Node] but one which will only stream updates for the node object
-// associated with the node we are currently running on.
-type LocalNodeResource resource.Resource[*slim_corev1.Node]
-
-func nodeResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*slim_corev1.Node], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*slim_corev1.NodeList](cs.Slim().CoreV1().Nodes()),
-		opts...,
-	)
-	return resource.New[*slim_corev1.Node](lc, lw), nil
-}
-
-// LocalCiliumNodeResource is a resource.Resource[*cilium_api_v2.Node] but one which will only stream updates for the
-// CiliumNode object associated with the node we are currently running on.
-type LocalCiliumNodeResource resource.Resource[*cilium_api_v2.CiliumNode]
-
-func ciliumNodeResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*cilium_api_v2.CiliumNode], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*cilium_api_v2.CiliumNodeList](cs.CiliumV2().CiliumNodes()),
-		opts...,
-	)
-	return resource.New[*cilium_api_v2.CiliumNode](lc, lw), nil
-}
-
-func namespaceResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*slim_corev1.Namespace], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*slim_corev1.NamespaceList](cs.Slim().CoreV1().Namespaces()),
-		opts...,
-	)
-	return resource.New[*slim_corev1.Namespace](lc, lw), nil
-}
-
-func lbIPPoolsResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*cilium_api_v2alpha1.CiliumLoadBalancerIPPool], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*cilium_api_v2alpha1.CiliumLoadBalancerIPPoolList](cs.CiliumV2alpha1().CiliumLoadBalancerIPPools()),
-		opts...,
-	)
-	return resource.New[*cilium_api_v2alpha1.CiliumLoadBalancerIPPool](lc, lw), nil
-}
-
-func ciliumIdentityResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*cilium_api_v2.CiliumIdentity], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*cilium_api_v2.CiliumIdentityList](cs.CiliumV2().CiliumIdentities()),
-		opts...,
-	)
-	return resource.New[*cilium_api_v2.CiliumIdentity](lc, lw), nil
-}
-
-func ciliumNetworkPolicyResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*cilium_api_v2.CiliumNetworkPolicy], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*cilium_api_v2.CiliumNetworkPolicyList](cs.CiliumV2().CiliumNetworkPolicies("")),
-		opts...,
-	)
-	return resource.New[*cilium_api_v2.CiliumNetworkPolicy](lc, lw), nil
-}
-
-func ciliumClusterwideNetworkPolicyResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*cilium_api_v2.CiliumClusterwideNetworkPolicy], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*cilium_api_v2.CiliumClusterwideNetworkPolicyList](cs.CiliumV2().CiliumClusterwideNetworkPolicies()),
-		opts...,
-	)
-	return resource.New[*cilium_api_v2.CiliumClusterwideNetworkPolicy](lc, lw), nil
-}
-
-func ciliumCIDRGroupResource(lc hive.Lifecycle, cs client.Clientset, opts ...func(*metav1.ListOptions)) (resource.Resource[*cilium_api_v2alpha1.CiliumCIDRGroup], error) {
-	if !cs.IsEnabled() {
-		return nil, nil
-	}
-	lw := utils.ListerWatcherWithModifiers(
-		utils.ListerWatcherFromTyped[*cilium_api_v2alpha1.CiliumCIDRGroupList](cs.CiliumV2alpha1().CiliumCIDRGroups()),
-		opts...,
-	)
-	return resource.New[*cilium_api_v2alpha1.CiliumCIDRGroup](lc, lw), nil
-}
